dev-requirements.txt
# Unit test requirements
flake8
html5lib
mock
pytest
pytest-cov
vcrpy
# Docs requirements
sphinx
sqlalchemy_schemadisplay
# Nice to have
flower
